clippy

+++ src/assets/mod.rs
@@ -13,7 +13,7 @@ const STRIDE: usize = 3;
 
 impl Texture {
     pub fn get_sample(&self, u: usize, v: usize) -> Color {
-        let index = ((v * self.width as usize) + u) * STRIDE;
+        let index = ((v * self.width) + u) * STRIDE;
         let slice = &self.data[index..index + STRIDE];
         Color::new(slice[0], slice[1], slice[2])
     }
